Step 13.4: Migrate RESTDataSource
