+++ modules/chats/index.ts
@@ -5,6 +5,7 @@ import commonModule from '../common';
 import usersModule from '../users';
 import { Message, Chat, pool } from '../../db';
 import { Resolvers } from '../../types/graphql';
+import { UnsplashApi } from './unsplash.api';
 
 const typeDefs = gql`
   type Message {
@@ -93,7 +94,7 @@ const resolvers: Resolvers = {
       return participant ? participant.name : null;
     },
 
-    async picture(chat, args, { currentUser, db, dataSources }) {
+    async picture(chat, args, { currentUser, db, unsplashApi }) {
       if (!currentUser) return null;
 
       const { rows } = await db.query(sql`
@@ -106,7 +107,7 @@ const resolvers: Resolvers = {
 
       return participant && participant.picture
         ? participant.picture
-        : dataSources.unsplashApi.getRandomPhoto();
+        : unsplashApi.getRandomPhoto();
     },
 
     async messages(chat, args, { db }) {
@@ -334,4 +335,9 @@ export default new GraphQLModule({
   typeDefs,
   resolvers,
   imports: () => [commonModule, usersModule],
+  context() {
+    return {
+      unsplashApi: new UnsplashApi(),
+    };
+  },
 });
