+++ package.json
@@ -39,6 +39,7 @@
     "apollo-datasource-rest": "0.5.0",
     "apollo-server-express": "2.6.0",
     "apollo-server-testing": "2.6.0",
+    "axios": "0.19.0",
     "bcrypt": "3.0.6",
     "body-parser": "1.19.0",
     "cookie": "0.4.0",
